# users/config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "N/A",
        "title": "SCIM",
        "description": "The SCIM API lets you manage users in your organization. You can then automate the provisioning of product licenses for these users, and they can use your company's Single Sign-On solution through an Identity Provider. The SCIM API lets you manage users in your organization. You can then automate the provisioning of product licenses for these users, and they can use your company's Single Sign-On solution through an Identity Provider."
    },
    "host": "pldgicrgsdfkmt6iq-mock.stoplight-proxy.io",
    "basePath": "/v1",
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/Groups/{groupKey}": {
            "parameters": [
                {
                    "name": "groupKey",
                    "in": "path",
                    "description": "The key of the group to query. The group must be in the organization domain",
                    "required": true,
                    "type": "integer",
                    "format": "int64"
                }
            ],
            "delete": {
                "operationId": "deleteGroup",
                "summary": "Delete Group",
                "tags": [
                    "Groups"
                ],
                "description": "Deletes a group from the organization (but not from the account). The group must be in the organization. This call requires the role ROLE_ORG_WRITE.",
                "responses": {
                    "200": {
                        "description": "The request has succeeded"
                    },
                    "401": {
                        "description": "Client is not sufficiently authorized or group does not exist in organization"
                    },
                    "403": {
                        "description": "Invalid token passed"
                    },
                    "502": {
                        "description": "Authentication or account gateway error occurred"
                    },
                    "504": {
                        "description": "Authentication or account gateway timeout occurred"
                    }
                }
            },
            "get": {
                "operationId": "getGroup",
                "summary": "Get Group",
                "tags": [
                    "Groups"
                ],
                "description": "Queries group details in the organization domain. This call requires the role ROLE_ORG_READ.",
                "responses": {
                    "200": {
                        "description": "The request has succeeded",
                        "schema": {
                            "$ref": "#/definitions/Group"
                        }
                    },
                    "401": {
                        "description": "Client is not sufficiently authorized or group does not exist in organization"
                    },
                    "403": {
                        "description": "Invalid token passed"
                    },
                    "502": {
                        "description": "Authentication or account gateway error occurred"
                    },
                    "504": {
                        "description": "Authentication or account gateway timeout occurred"
                    }
                }
            },
            "patch": {
                "operationId": "updateGroup",
                "summary": "Update Group",
                "tags": [
                    "Groups"
                ],
                "description": "Updates one or more values of an existing group without sending the full definition. Member groups and member users must be in the organization. This call requires the role ROLE_ORG_WRITE.",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/GroupDefinition"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The group has been updated",
                        "schema": {
                            "$ref": "#/definitions/Group"
                        }
                    },
                    "400": {
                        "description": "Possible reasons are:<br />The displayName field is not set<br />The displayName field is malformed<br />The displayName field exceeds 128 characters<br />The members array exceeds 100 elements<br />"
                    },
                    "401": {
                        "description": "Client is not sufficiently authorized or group does not exist in organization"
                    },
                    "403": {
                        "description": "Invalid token passed"
                    },
                    "409": {
                        "description": "The displayName is already in use within the organization"
                    },
                    "502": {
                        "description": "Authentication or account gateway error occurred"
                    },
                    "504": {
                        "description": "Authentication or account gateway timeout occurred"
                    }
                }
            },
            "put": {
                "operationId": "replaceGroup",
                "summary": "Replace Group",
                "tags": [
                    "Groups"
                ],
                "description": "Updates an existing group. The request must include the full group definition. To modify one or more values without sending the full definition, see \"Update Group\". Member groups and member users must be in the organization. This call requires the role ROLE_ORG_WRITE.",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/GroupDefinition"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The group has been replaced.",
                        "schema": {
                            "$ref": "#/definitions/Group"
                        }
                    },
                    "400": {
                        "description": "Possible reasons are:<br />The displayName field is not set<br />The displayName field is malformed<br />The displayName field exceeds 128 characters<br />The members array exceeds 100 elements<br />"
                    },
                    "401": {
                        "description": "Client is not sufficiently authorized or group does not exist in organization"
                    },
                    "403": {
                        "description": "Invalid token passed"
                    },
                    "409": {
                        "description": "The displayName is already in use within the organization"
                    },
                    "502": {
                        "description": "Authentication or account gateway error occurred"
                    },
                    "504": {
                        "description": "Authentication or account gateway timeout occurred"
                    }
                }
            }
        },
...
        "/Users/me": {
            "put": {
                "operationId": "replaceMe",
                "summary": "Replace Current User",
                "tags": [
                    "Users"
                ],
                "description": "Changes the current authenticated user's displayName, locale, timezone, username and password. The request must include the full user definition (to modify one or more values without sending the full definition, see Update User). The replaced user email domain must be an existing organization email domain.",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/UserDefinition"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The current user has been replaced.",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "One of the following requirements is not met:<br />The userName field is required.<br />The userName field must be non-empty.<br />The userName field cannot exceed 128 characters.<br />The locale field must be composed of 1 or 2 parts.<br />The locale language sub-field must be ISO-639.<br />The locale country sub-field must be ISO-3166.<br />The timezone field must be a valid timezone.<br />The givenName field must not exceed 60 characters.<br />The familyName field must not exceed 60 characters."
                    },
                    "401": {
                        "description": "Client is not sufficiently authorized"
                    },
                    "403": {
                        "description": "Invalid token passed"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "409": {
                        "description": "Email address conflict"
                    },
                    "502": {
                        "description": "Authentication or account gateway error occurred"
                    },
                    "504": {
                        "description": "Authentication or account gateway timeout occurred"
                    }
                }
            },
            "patch": {
                "operationId": "updateMe",
                "summary": "Update Current User",
                "tags": [
                    "Users"
                ],
                "description": "Changes a limited set (or all if you choose) of the current authenticated user's data. The updated user email domain must be an existing organization email domain.",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/UserDefinition"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The current user has been updated.",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "One of the following requirements is not met:<br />The userName field is required.<br />The userName field must be non-empty.<br />The userName field cannot exceed 128 characters.<br />The locale field must be composed of 1 or 2 parts.<br />The locale language sub-field must be ISO-639.<br />The locale country sub-field must be ISO-3166.<br />The timezone field must be a valid timezone.<br />The givenName field must not exceed 60 characters.<br />The familyName field must not exceed 60 characters."
                    },
                    "401": {
                        "description": "Client is not sufficiently authorized"
                    },
                    "403": {
                        "description": "Invalid token passed"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "409": {
                        "description": "Email address conflict"
                    },
                    "502": {
                        "description": "Authentication or account gateway error occurred"
                    },
                    "504": {
                        "description": "Authentication or account gateway timeout occurred"
                    }
                }
            },
            "get": {
                "operationId": "getMe",
                "summary": "Get Current User",
                "tags": [
                    "Users"
                ],
                "description": "Queries the identity of the current authenticated user.",
                "responses": {
                    "200": {
                        "description": "The request has succeeded.",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "401": {
                        "description": "Client is not sufficiently authorized"
                    },
                    "403": {
                        "description": "Invalid token passed"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "502": {
                        "description": "Authentication or account gateway error occurred"
                    },
                    "504": {
                        "description": "Authentication or account gateway timeout occurred"
                    }
                }
            }
        },
...
    },
...
    "definitions": {
        "UserDefinition": {
            "title": "UserDefinition",
            "description": "Describes a new user",
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string",
                    "description": "The user's display name",
                    "x-faker": "internet.userName"
                },
                "locale": {
                    "description": "The user's language settings",
                    "type": "string"
                },
                "name": {
                    "$ref": "#/definitions/UserFullName"
                },
                "timezone": {
                    "description": "The user's time zone",
                    "type": "string"
                },
                "userName": {
                    "description": "The user's username, usually their email address",
                    "type": "string",
                    "x-faker": "internet.userName"
                },
                "emails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Email"
                    }
                },
                "phoneNumbers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/PhoneNumber"
                    }
                },
                "schemas": {
                    "type": "array",
                    "default": [],
                    "uniqueItems": true,
                    "items": {
                        "type": "string",
                        "enum": [
                            "urn:scim:schemas:core:1.0",
                            "urn:scim:schemas:extension:italia:1.0"
                        ]
                    }
                },
                "urn:scim:schemas:extension:italia:1.0": {
                    "type": "object",
                    "required": [
                        "fiscalNumber"
                    ],
                    "properties": {
                        "fiscalNumber": {
                            "type": "string"
                        },
                        "optOut": {
                            "type": "array",
                            "uniqueItems": true,
                            "items": {
                                "type": "object"
                            }
                        }
                    }
                }
            },
            "required": [
                "displayName",
                "name",
                "emails",
                "schemas",
                "urn:scim:schemas:extension:italia:1.0"
            ]
        },
...
        "GroupMetadata": {
            "title": "GroupMetadata",
            "description": "Group metadata",
            "properties": {
                "created": {
                    "description": "The date and time the group was created",
                    "format": "date-time",
                    "type": "string"
                },
                "location": {
                    "description": "The URL where the group can be accessed",
                    "type": "string"
                }
            },
            "required": [
                "created",
                "location"
            ]
        },
...
        "Group": {
            "title": "Group",
            "description": "Describes a group belonging to an organization",
            "properties": {
                "displayName": {
                    "description": "The group's display name",
                    "type": "string"
                },
                "id": {
                    "description": "The group's unique key",
                    "type": "string"
                },
                "members": {
                    "description": "An array of members",
                    "items": {
                        "$ref": "#/definitions/Member"
                    },
                    "type": "array"
                },
                "meta": {
                    "$ref": "#/definitions/GroupMetadata",
                    "description": "Group metadata"
                }
            },
            "required": [
                "meta",
                "id",
                "displayName"
            ]
        },
        "Member": {
            "title": "Member",
            "description": "A member of a group. This can be a group or an user",
            "properties": {
                "type": {
                    "description": "The type of member",
                    "enum": [
                        "group",
                        "user"
                    ],
                    "type": "string"
                },
                "value": {
                    "description": "Depending on the member type this is a groupKey or an userKey",
                    "type": "string"
                }
            },
            "required": [
                "type",
                "value"
            ]
        },
        "GroupDefinition": {
            "title": "GroupDefinition",
            "description": "Describes the group to create",
            "properties": {
                "displayName": {
                    "description": "The group's display name",
                    "type": "string"
                },
                "members": {
                    "description": "An array of members",
                    "items": {
                        "$ref": "#/definitions/Member"
                    },
                    "type": "array"
                }
            },
            "required": [
                "displayName"
            ]
        },
...
        "UserMetadata": {
            "title": "UserMetadata",
            "description": "User metadata",
            "properties": {
                "created": {
                    "description": "The date and time the user was created",
                    "format": "date-time",
                    "type": "string"
                },
                "location": {
                    "description": "The URL where the user can be accessed",
                    "type": "string"
                }
            },
            "required": [
                "created",
                "location"
            ]
        },
        "UserFullName": {
            "title": "UserFullName",
            "description": "The full name of the user",
            "properties": {
                "familyName": {
                    "description": "The user's surname",
                    "type": "string",
                    "x-faker": "name.lastName"
                },
                "givenName": {
                    "description": "The user's first name",
                    "type": "string",
                    "x-faker": "name.firstName"
                }
            },
            "required": [
                "familyName",
                "givenName"
            ]
        },
        "User": {
            "title": "User",
            "description": "Describes a user belonging to an organization",
            "allOf": [
                {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "string",
                            "description": "The user's unique key"
                        },
                        "meta": {
                            "$ref": "#/definitions/UserMetadata"
                        }
                    },
                    "required": [
                        "id"
                    ]
                },
                {
                    "$ref": "#/definitions/UserDefinition"
                }
            ]
        },
...
        "Email": {
            "title": "Email",
            "type": "object",
            "properties": {
                "type": {
                    "type": "string",
                    "x-faker": "lorem.word"
                },
                "primary": {
                    "type": "boolean"
                },
                "value": {
                    "type": "string",
                    "format": "email",
                    "x-faker": "internet.email"
                }
            },
            "required": [
                "value"
            ]
        },
        "PhoneNumber": {
            "title": "PhoneNumber",
            "type": "object",
            "properties": {
                "value": {
                    "type": "string",
                    "x-faker": "phone.phoneNumber"
                },
                "type": {
                    "type": "string",
                    "x-faker": "lorem.word"
                }
            },
            "required": [
                "value"
            ],
            "example": {
                "value": "1-047-149-5200 x41058",
                "type": "ut"
            }
        }
    }
}